+++ database.go
@@ -1,0 +1,27 @@
+package main
+
+import (
+	"database/sql"
+	"fmt"
+	"log"
+
+	_ "github.com/lib/pq"
+)
+
+func ConnectDB() {
+	connectionString := fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=%s sslmode=disable",
+		EnvConfig.DbHost, EnvConfig.DbPort, EnvConfig.DbUser, EnvConfig.DbPassword, EnvConfig.DbName)
+
+	db, err := sql.Open("postgres", connectionString)
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer db.Close()
+
+	err = db.Ping()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	log.Println("Successfully connected to PostgreSQL")
+}